Creating delete bundle for non-redeployable encass
